DATAGRAHPH-1120 - Align package structure with JPA and Mongo; be clear of filters being private API.
