--- spring-data-neo4j/src/main/java/org/springframework/data/neo4j/repository/query/derived/builder/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Builder classes to create Neo4j-OGM filters.
- */
-package org.springframework.data.neo4j.repository.query.derived.builder;
