--- spring-data-neo4j/src/main/java/org/springframework/data/neo4j/repository/query/derived/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Derived query creation.
- */
-package org.springframework.data.neo4j.repository.query.derived;
