+++ spring-data-neo4j/src/main/java/org/springframework/data/neo4j/repository/query/filter/ContainsComparisonBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c)  [2011-2017] "Pivotal Software, Inc." / "Neo Technology" / "Graph Aware Ltd."
+ * Copyright (c)  [2011-2018] "Pivotal Software, Inc." / "Neo Technology" / "Graph Aware Ltd."
  *
  * This product is licensed to you under the Apache License, Version 2.0 (the "License").
  * You may not use this product except in compliance with the License.
@@ -11,7 +11,7 @@
  *
  */
 
-package org.springframework.data.neo4j.repository.query.derived.builder;
+package org.springframework.data.neo4j.repository.query.filter;
 
 import java.util.Collections;
 import java.util.List;
@@ -28,9 +28,9 @@ import org.springframework.data.repository.query.parser.Part;
  *
  * @author Gerrit Meier
  */
-public class ContainsComparisonBuilder extends FilterBuilder {
+class ContainsComparisonBuilder extends FilterBuilder {
 
-	public ContainsComparisonBuilder(Part part, BooleanOperator booleanOperator, Class<?> entityType) {
+	ContainsComparisonBuilder(Part part, BooleanOperator booleanOperator, Class<?> entityType) {
 		super(part, booleanOperator, entityType);
 	}
