+++ spring-data-neo4j/src/main/java/org/springframework/data/neo4j/repository/query/filter/FilterBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c)  [2011-2016] "Pivotal Software, Inc." / "Neo Technology" / "Graph Aware Ltd."
+ * Copyright (c)  [2011-2018] "Pivotal Software, Inc." / "Neo Technology" / "Graph Aware Ltd."
  *
  * This product is licensed to you under the Apache License, Version 2.0 (the "License").
  * You may not use this product except in compliance with the License.
@@ -11,7 +11,7 @@
  *
  */
 
-package org.springframework.data.neo4j.repository.query.derived.builder;
+package org.springframework.data.neo4j.repository.query.filter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,12 +26,35 @@ import org.springframework.data.repository.query.parser.Part;
  * @author Jasper Blues
  * @author Nicolas Mervaillie
  * @author Gerrit Meier
+ * @author Michael J. Simons
  */
 public abstract class FilterBuilder {
 
 	protected Part part;
 	protected BooleanOperator booleanOperator;
 	protected Class<?> entityType;
+
+	public static FilterBuilder forPartAndEntity(Part part, Class<?> entityType, BooleanOperator booleanOperator) {
+		switch (part.getType()) {
+			case NEAR:
+				return new DistanceComparisonBuilder(part, booleanOperator, entityType);
+			case BETWEEN:
+				return new BetweenComparisonBuilder(part, booleanOperator, entityType);
+			case NOT_CONTAINING:
+			case CONTAINING:
+				return resolveMatchingContainsFilterBuilder(part, entityType, booleanOperator);
+			case IS_NULL:
+			case IS_NOT_NULL:
+				return new IsNullFilterBuilder(part, booleanOperator, entityType);
+			case EXISTS:
+				return new ExistsFilterBuilder(part, booleanOperator, entityType);
+			case TRUE:
+			case FALSE:
+				return new BooleanComparisonBuilder(part, booleanOperator, entityType);
+			default:
+				return new PropertyComparisonBuilder(part, booleanOperator, entityType);
+		}
+	}
 
 	FilterBuilder(Part part, BooleanOperator booleanOperator, Class<?> entityType) {
 		this.part = part;
@@ -49,7 +72,7 @@ public abstract class FilterBuilder {
 		return part.getProperty().getSegment();
 	}
 
-	void setNestedAttributes(Part part, Filter filter) {
+	protected void setNestedAttributes(Part part, Filter filter) {
 		List<Filter.NestedPathSegment> segments = new ArrayList<>();
 		PropertyPath property = part.getProperty();
 		if (property.hasNext()) {
@@ -74,4 +97,12 @@ public abstract class FilterBuilder {
 		return segments;
 	}
 
+	private static FilterBuilder resolveMatchingContainsFilterBuilder(Part part, Class<?> entityType,
+			BooleanOperator booleanOperator) {
+		boolean usePropertyComparison = !part.getProperty().getTypeInformation().isCollectionLike();
+		if (usePropertyComparison) {
+			return new PropertyComparisonBuilder(part, booleanOperator, entityType);
+		}
+		return new ContainsComparisonBuilder(part, booleanOperator, entityType);
+	}
 }
