+++ spring-data-neo4j/src/main/java/org/springframework/data/neo4j/repository/query/filter/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Support for creating Cypher queries through Neo4j-OGMs filter mechanism. Not to be meant as part of an public api.
+ */
+package org.springframework.data.neo4j.repository.query.filter;
